requirements.txt
streamlit
clickhouse_connect
python-dotenv
ollama